4090/DariaVazhenina/llm-finetune/requirements.txt
llm-foundry
protobuf==3.20.0
bitsandbytes
transformers
accelerate
huggingface-hub==0.17.3

transformers==4.*
huggingface-hub==0.*
emoji==1.*
demoji==1.*
fugashi==1.*
nagisa==0.2.*
pytorch-lightning==1.*
torchmetrics==0.7.*
unidic-lite==1.*
ipadic==1.*
pyarrow==7.*
sentencepiece==0.1.*
boto3>=1.22.9
datasets==2.6.*
gensim==4.2.*
loguru==0.6.*
lap==0.4.*
yacs==0.1.*
thop==0.1.*
swifter==1.3.*
lightly==1.4.*
gptq==0.0.*
auto-gptq==0.3.*
